Halo
Finally the publishing of Halo aka BitHalo, BlackHalo, BitBay to github.
The first commit is BitMHalo which handles all the decentralized
markets, IMAP and messaging. It can not run without the Halo obfuscated
package however the functionality works when Halo is running. It uses a
locking file instead of RPC calls to avoid dropping commands. There are
various ways to improve this eventually.